+++ src/app/organization/bulk-export/bulk-export.component.ts
@@ -1,0 +1,18 @@
+/** Angular Imports */
+import { Component, OnInit } from '@angular/core';
+
+/**
+ * Bulk Export component.
+ */
+@Component({
+  selector: 'mifosx-bulk-export',
+  templateUrl: './bulk-export.component.html',
+  styleUrls: ['./bulk-export.component.scss']
+})
+export class BulkExportComponent implements OnInit {
+
+  constructor() {}
+
+  ngOnInit() {}
+
+}